Refactor to support multi-communication library--ClientEngine
