+++ ams/src/main/resources/com/flying/ams/ams-config.xml
@@ -11,7 +11,7 @@
         <property name="scheduler" ref="scheduler"/>
     </bean>
 
-    <bean id="accountServerEngine" class="com.flying.framework.messaging.engine.impl.ZMQUCServerEngine">
+    <bean id="accountServerEngine" class="com.flying.framework.messaging.engine.impl.zmq.ZMQUCServerEngine">
         <constructor-arg name="listenEndpoint" ref="listenEndpoint"/>
         <property name="msgEventListener" ref="amsMsgListener"/>
         <property name="listenEndpoint">
@@ -58,7 +58,7 @@
     <!--<constructor-arg name="msgConverter" ref="monitorMsgConverter"/>-->
     <!--</bean>-->
 
-    <bean id="broadcastEngine" class="com.flying.framework.messaging.engine.impl.BroadcastClientEngine" init-method="start" destroy-method="stop">
+    <bean id="broadcastEngine" class="com.flying.framework.messaging.engine.impl.jdk.BroadcastClientEngine" init-method="start" destroy-method="stop">
         <constructor-arg name="endpoints" ref="broadcastEndpoints"/>
     </bean>
     <bean id="monitorMsgConverter" class="com.flying.monitor.msg.converter.MonitorMsgConverter"/>
