+++ framework/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     testCompile group: 'org.hamcrest', name: 'hamcrest-core', version: '1.3'
 
     compile group: 'org.zeromq', name: 'jzmq', version: '3.1.1-SNAPSHOT'
+    compile group: 'io.netty', name: 'netty-all', version: '4.1.9.Final'
     compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.25'
     compile group: 'com.google.guava', name: 'guava', version: '18.0'
