+++ framework/src/main/java/com/flying/framework/messaging/engine/IAsyncClientEngineConfig.java
@@ -1,0 +1,27 @@
+/*
+ Created by Walker on 2017/4/9.
+ Revision History:
+ Date          Who              Version      What
+ 2017/4/9      Walker           0.1.0        Created. 
+*/
+package com.flying.framework.messaging.engine;
+
+import com.flying.framework.messaging.endpoint.IEndpoint;
+import com.flying.framework.messaging.event.IMsgEventListener;
+
+import java.util.List;
+import java.util.concurrent.Executor;
+
+public interface IAsyncClientEngineConfig {
+    List<IEndpoint> getEndpoints();
+
+    void setEndpoints(List<IEndpoint> endpoints);
+
+    Executor getExecutor();
+
+    void setExecutor(Executor executor);
+
+    IMsgEventListener getMsgEventListener();
+
+    void setMsgEventListener(IMsgEventListener msgEventListener);
+}
