+++ monitor/src/main/resources/com/flying/monitor/monitor-broadcast-config.xml
@@ -22,7 +22,7 @@
         <property name="delay" value="5000"/>
         <property name="initialDelay" value="3000"/>
     </bean>
-    <bean id="monitorServerEngine" class="com.flying.framework.messaging.engine.impl.ZMQBCServerEngine">
+    <bean id="monitorServerEngine" class="com.flying.framework.messaging.engine.impl.zmq.ZMQBCServerEngine">
         <constructor-arg name="listenEndpoint" ref="monitorListenEndPoint"/>
         <property name="msgEventListener" ref="monitorMsgListener"/>
     </bean>
@@ -65,7 +65,7 @@
     </bean>
     <bean id="monitorServerService" class="com.flying.monitor.service.server.MonitorServerService"/>
 
-    <bean id="broadcastEngine" class="com.flying.framework.messaging.engine.impl.BroadcastClientEngine" init-method="start" destroy-method="stop">
+    <bean id="broadcastEngine" class="com.flying.framework.messaging.engine.impl.jdk.BroadcastClientEngine" init-method="start" destroy-method="stop">
         <constructor-arg name="endpoints" ref="broadcastEndpoints"/>
     </bean>
     <bean id="monitorMsgConverter" class="com.flying.monitor.msg.converter.MonitorMsgConverter"/>
