+++ monitor/src/main/resources/com/flying/monitor/monitor-unicast-config.xml
@@ -22,7 +22,7 @@
         <property name="delay" value="5000"/>
         <property name="initialDelay" value="3000"/>
     </bean>
-    <bean id="ucMonitorServerEngine" class="com.flying.framework.messaging.engine.impl.ZMQUCServerEngine">
+    <bean id="ucMonitorServerEngine" class="com.flying.framework.messaging.engine.impl.zmq.ZMQUCServerEngine">
         <constructor-arg name="listenEndpoint" ref="monitorListenEndpoint"/>
         <property name="msgEventListener" ref="monitorMsgListener"/>
     </bean>
