+++ simulator/src/main/java/com/patrol/simulator/PingerMain.java
@@ -10,7 +10,7 @@ package com.patrol.simulator;
 import com.flying.framework.messaging.endpoint.impl.Endpoint;
 import com.flying.framework.messaging.endpoint.IEndpoint;
 import com.flying.framework.messaging.engine.IPinger;
-import com.flying.framework.messaging.engine.impl.ZMQPinger;
+import com.flying.framework.messaging.engine.impl.zmq.ZMQPinger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
